Refactor OOBE VMs
