+++ Natsurainko.FluentLauncher/ViewModels/OOBE/OOBEViewModel.cs
@@ -3,28 +3,28 @@ using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Media.Animation;
 using Natsurainko.FluentLauncher.Services.Accounts;
 using Natsurainko.FluentLauncher.Services.Launch;
 using Natsurainko.FluentLauncher.Services.Settings;
 using Natsurainko.FluentLauncher.Services.UI;
 using Natsurainko.FluentLauncher.Services.UI.Messaging;
 using Natsurainko.FluentLauncher.Services.UI.Navigation;
+using Natsurainko.FluentLauncher.Utils;
+using Natsurainko.FluentLauncher.Views.Common;
+using Nrk.FluentCore.Classes.Datas.Authenticate;
+using Nrk.FluentCore.Utils;
 using System;
-using System.Windows.Forms.VisualStyles;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+using Windows.Storage.Pickers;
 
 namespace Natsurainko.FluentLauncher.ViewModels.OOBE;
 
-internal partial class OOBEViewModel : ObservableObject, INavigationAware, ISettingsViewModel
+internal partial class OOBEViewModel : ObservableRecipient, INavigationAware, ISettingsViewModel, IRecipient<ActiveAccountChangedMessage>
 {
-    #region ObservableProperties
-
-    [ObservableProperty]
-    [NotifyCanExecuteChangedFor(nameof(BackCommand))]
-    int currentPageIndex;
-
-    #endregion
-
     #region Dependencies
 
     [SettingsProvider]
@@ -36,17 +36,6 @@ internal partial class OOBEViewModel : ObservableObject, INavigationAware, ISett
     private readonly AccountService _accountService;
 
     #endregion
-
-    private static readonly string[] OOBEPageKeys =
-    {
-        "OOBELanguagePage",
-        "OOBEBasicPage",
-        "OOBEAccountPage",
-        "OOBEGetStartedPage"
-    };
-
-
-    private Type NextPage;
 
     public OOBEViewModel(
         INavigationService navigationService,
@@ -61,6 +50,11 @@ internal partial class OOBEViewModel : ObservableObject, INavigationAware, ISett
         _notificationService = notificationService;
         _accountService = accountService;
 
+        // Init accounts
+        Accounts = accountService.Accounts;
+        ActiveAccount = accountService.ActiveAccount;
+        IsActive = true; // Enable message recipient
+
         (this as ISettingsViewModel).InitializeSettings();
 
         //WeakReferenceMessenger.Default.Register<GuideNavigationMessage>(this, (r, m) =>
@@ -72,30 +66,56 @@ internal partial class OOBEViewModel : ObservableObject, INavigationAware, ISett
 
     #region Navigation
 
+    [ObservableProperty]
+    [NotifyCanExecuteChangedFor(nameof(BackCommand))]
+    int currentPageIndex;
+
+    private static readonly string[] OOBEPageKeys =
+    {
+        "OOBELanguagePage",
+        "OOBEBasicPage",
+        "OOBEAccountPage",
+        "OOBEGetStartedPage"
+    };
+
     void INavigationAware.OnNavigatedTo(object parameter)
     {
         _navigationService.NavigateTo("OOBELanguagePage"); // Default page
     }
 
     [RelayCommand(CanExecute = nameof(CanNext))]
-    public void Next(Frame frame)
+    public void Next()
     {
         CurrentPageIndex++;
         _navigationService.NavigateTo(OOBEPageKeys[CurrentPageIndex]);
+        BackCommand.NotifyCanExecuteChanged();
+        NextCommand.NotifyCanExecuteChanged();
     }
 
-    bool CanNext()
+    bool CanNext() => CurrentPageIndex switch
     {
-        // TODO: validation for NextCommand
-
-        return true;
-    }
+        // Language page
+        0 => Languages.Contains(CurrentLanguage),
+        // Basic page
+        1 => !string.IsNullOrEmpty(ActiveMinecraftFolder) &&
+                                    !string.IsNullOrEmpty(ActiveJava) &&
+                                    Directory.Exists(ActiveMinecraftFolder) &&
+                                    File.Exists(ActiveJava),
+        // Account page
+        2 => ActiveAccount is not null,
+        // Get started page
+        3 => false,
+        // Default
+        _ => false,
+    };
 
     [RelayCommand(CanExecute = nameof(CanBack))]
-    public void Back(Frame frame)
+    public void Back()
     {
         _navigationService.GoBack();
         CurrentPageIndex--;
+        BackCommand.NotifyCanExecuteChanged();
+        NextCommand.NotifyCanExecuteChanged();
     }
 
     bool CanBack()
@@ -108,4 +128,146 @@ internal partial class OOBEViewModel : ObservableObject, INavigationAware, ISett
     }
 
     #endregion
+
+    #region Language
+
+    [ObservableProperty]
+    [NotifyCanExecuteChangedFor(nameof(NextCommand))]
+    [BindToSetting(Path = nameof(SettingsService.CurrentLanguage))]
+    private string currentLanguage;
+
+    public List<string> Languages { get; } = ResourceUtils.Languages;
+
+    partial void OnCurrentLanguageChanged(string oldValue, string newValue)
+    {
+        if (Languages.Contains(CurrentLanguage) && oldValue is not null) // oldValue is null at startup
+            ResourceUtils.ApplyLanguage(CurrentLanguage);
+    }
+
+    #endregion
+
+    #region Minecraft folder
+
+    [BindToSetting(Path = nameof(SettingsService.MinecraftFolders))]
+    public ObservableCollection<string> MinecraftFolders { get; private set; }
+
+    [ObservableProperty]
+    [NotifyCanExecuteChangedFor(nameof(NextCommand))]
+    [BindToSetting(Path = nameof(SettingsService.ActiveMinecraftFolder))]
+    private string activeMinecraftFolder;
+
+    [RelayCommand]
+    public Task BrowserFolder() => Task.Run(async () =>
+    {
+        var folderPicker = new FolderPicker();
+
+        var hwnd = WinRT.Interop.WindowNative.GetWindowHandle(App.MainWindow);
+        WinRT.Interop.InitializeWithWindow.Initialize(folderPicker, hwnd);
+
+        folderPicker.FileTypeFilter.Add("*");
+        var folder = await folderPicker.PickSingleFolderAsync();
+
+        if (folder != null)
+            App.DispatcherQueue.TryEnqueue(() =>
+            {
+                if (MinecraftFolders.Contains(folder.Path))
+                {
+                    _notificationService.NotifyMessage("Failed to add the folder", "This folder already exists", icon: "\uF89A");
+                    return;
+                }
+
+                MinecraftFolders.Add(folder.Path);
+                _gameService.ActivateMinecraftFolder(folder.Path);
+            });
+    });
+
+    #endregion
+
+    #region Java
+
+    [BindToSetting(Path = nameof(SettingsService.Javas))]
+    public ObservableCollection<string> Javas { get; private set; }
+
+    [ObservableProperty]
+    [BindToSetting(Path = nameof(SettingsService.ActiveJava))]
+    [NotifyCanExecuteChangedFor(nameof(NextCommand))]
+    private string activeJava;
+
+    [ObservableProperty]
+    private bool javaDropDownOpen;
+
+    [RelayCommand]
+    public Task BrowserJava() => Task.Run(async () =>
+    {
+        var filePicker = new FileOpenPicker();
+
+        var hwnd = WinRT.Interop.WindowNative.GetWindowHandle(App.MainWindow);
+        WinRT.Interop.InitializeWithWindow.Initialize(filePicker, hwnd);
+
+        filePicker.FileTypeFilter.Add(".exe");
+        var file = await filePicker.PickSingleFileAsync();
+
+        if (file != null)
+            App.DispatcherQueue.TryEnqueue(() =>
+            {
+                Javas.Add(file.Path);
+                OnPropertyChanged(nameof(Javas));
+
+                ActiveJava = file.Path;
+            });
+    });
+
+    [RelayCommand]
+    public void SearchJava()
+    {
+        foreach (var java in JavaUtils.SearchJava())
+            if (!Javas.Contains(java))
+                Javas.Add(java);
+
+        ActiveJava = Javas.Any() ? Javas[0] : null;
+
+        OnPropertyChanged(nameof(Javas));
+
+        JavaDropDownOpen = true;
+        _notificationService.NotifyWithoutContent("Added the search Java to the runtime list", icon: "\uE73E");
+    }
+
+    #endregion
+
+    #region Account
+
+    public ReadOnlyObservableCollection<Account> Accounts { get; init; }
+
+    [ObservableProperty]
+    [NotifyCanExecuteChangedFor(nameof(NextCommand))]
+    private Account activeAccount;
+
+    bool processingActiveAccountChangedMessage = false;
+
+    public void Receive(ActiveAccountChangedMessage message)
+    {
+        processingActiveAccountChangedMessage = true;
+        ActiveAccount = message.Value;
+        processingActiveAccountChangedMessage = false;
+    }
+
+    partial void OnActiveAccountChanged(Account value)
+    {
+        if (!processingActiveAccountChangedMessage)
+            _accountService.Activate(value);
+    }
+
+    [RelayCommand]
+    public void Login(Button parameter) 
+        => _ = new AuthenticationWizardDialog { XamlRoot = parameter.XamlRoot }.ShowAsync();
+
+    #endregion
+
+    [RelayCommand]
+    public void Start()
+    {
+        _navigationService.Parent?.NavigateTo("ShellPage");
+        _settings.FinishGuide = true;
+    }
+
 }
